requirements.txt
Flask>=2.3
flask-cors
flask-limiter
flask-pydantic
pydantic
PyJWT
python-dotenv
psycopg[binary]
structlog
pydantic-settings

